% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Function to retrieve data from the monthly output file created by NUCOMBog}
\usage{
getData(WD, type)
}
\arguments{
\item{WD}{Working Directory}

\item{type}{Data to be retrieved, integrated are "NEE" and "WTD"}
}
\description{
Function to retrieve data from the monthly output file created by NUCOMBog
}
\examples{
\dontrun{
getData(WD="~/nucom",type = c("NEE","WTD"))
}
}
\author{
Jeroen Pullens
}

