\name{make_filenames}
\alias{make_filenames}
\title{make_filenames to run the model, the model needs this file to be able to run}
\usage{
make_filenames(WD)
}
\arguments{
\item{WD} {"Working directory"}

}
\value{
Makes a file to start the model
}
\description{
Makes a file to start the model
}
\examples{
make_filenames(WD)
}



