% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_file_loop.R
\name{make_param_file}
\alias{make_param_file}
\title{make parameter list}
\usage{
make_param_file(WD, par)
}
\arguments{
\item{WD}{Working directory}

\item{par}{possibility to add own parameters, BUT is has to in a datafram format with colum names: "names" and "values", see example.}
}
\description{
Code to make a parameter list. The parameter list contains all the parameters for NUCOMBog. It is incorporated in the runnucom() function. By default it is set to simulate the LVM of Heijmans et al. 2008
}
\examples{
\dontrun{
names<-c("CO2ref","CritNClvms")
values<-c(300,0.1)
par<-data.frame(names,values)
make_param_file("~/nucom/",par)
}
}
\author{
Jeroen Pullens
}
\keyword{NUCOMBog}

