\name{make_param_file_complete}
\alias{make_param_file_complete}
\title{make_param_file_complete}
\usage{
make_param_file_complete(WD,x)
}
\arguments{
\item{WD} {"Working directory"}
\item{x} {"Parameter valus (list of 198 values)"}

}
\value{
Makes a parameter file to start the model
}
\description{
Makes a parameter file to start the model
}
\examples{
make_param_file_complete(WD,x)
}
