% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_param_file_subset.r
\name{make_param_file_subset}
\alias{make_param_file_subset}
\title{make_param_file_subset}
\usage{
make_param_file_subset(WD, x)
}
\arguments{
\item{WD}{Working Directory}

\item{x}{values for the parameters}
}
\description{
make_param_file_subset
}
\examples{
\dontrun{
make_param_file_subset(WD,c(1:26))
}
}
\author{
Jeroen Pullens
}

