% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUCOMBog.R
\docType{package}
\name{NUCOMBog}
\alias{NUCOMBog}
\alias{NUCOMBog-package}
\title{NUCOMBog}
\source{
The executable and the source code of the model can downloaded from \url{https://github.com/jeroenpullens/NUCOMBog}.
}
\description{
The NUCOMBog R package provides an interface to the NUCOMBog model in R.

The package simulates the dynamics of five plant functional types (PFTs): graminoids, ericaceous shrubs and three groups of \emph{Sphagnum} mosses (lawn, hollow and hummock mosses) on a monthly time step. The R package also calculates the monthly heterotrophic respiration and therefore the net ecosystem exchange can be calculated.
The package provides a user-friendly tool that allows simulating peatlands over years/decades, under different management strategies and climate change scenarios.


For details on how to use the package, go to the help files of the functions.

This work was supported by a STSM grant to JWM Pullens from COST Action FP1304 (Profound, \url{http://cost-profound.eu/site/})
}
\author{
JWM Pullens
}

