% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runnucom.r
\name{runnucom}
\alias{runnucom}
\title{Run NUCOMBog}
\usage{
runnucom(setup, parameters = NULL)
}
\arguments{
\item{setup}{Add here all the essential information to run the model.}

\item{parameters}{The parameters which are used in the model. If no parameter values are given the default values will be used. The parameters have to have the format of a dataframe with colum names: "names" and "values", see example.}
}
\description{
Code to run NUCOMBog.
}
\examples{
\dontrun{
Single core:
test_setup_singlecore <- setup_NUCOM(mainDir="/home/jeroen/test_package/",
                                     climate="clim_1999-2013_measured.txt",
                                     environment="Env_Mer_Bleue_1999_2013.txt",
                                     inival="Inival_Mer_Bleue.txt",
                                     start=1999,
                                     end=2013,
                                     type=c("NEE","WTD"),
                                     parallel=F)


names<-c("CO2ref","gram_Beta","sd1","sd2")
parameters<-list()
nparvector<-50
for (i in 1:nparvector){
parameters[[i]]<-data.frame(names,
                           values=runif(n=length(names),
                           min=c(300,0.1,0.01,0.01),
                           max=c(500,1,1,1)))
                           }

runnucom(setup = test_setup_singlecore,parameters=parameters)

## without predefined parameters:
runnucom(setup = test_setup_singlecore,parameters=NULL)
}
}
\author{
JWM Pullens
}

