% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runnucomParallel.r
\name{runnucom_parallel}
\alias{runnucom_parallel}
\title{Run NUCOMBog parallel}
\usage{
runnucom_parallel(Setup, mainDir, clustertype, numCores)
}
\arguments{
\item{Setup}{list of:
       $WD = Working directory
       $climate = name of climate file
       $environment = name of environment file
       $inival = name of initial values file
       $start  = start year
       $end = end year
       $par = data.frame(names,values)
       $type = output}

\item{mainDir}{main directory where the files are stored}

\item{clustertype}{Clustertype: "MPI" is cluster and personal computer is "SOCK"}

\item{numCores}{Number of Cores}
}
\description{
Code to run NUCOMBog parallel, but first run setupParallel
}
\examples{
\dontrun{
!!test_run is from the function setupParallel!!
parallel_output<-runnucom_parallel(Setup=test_setup,mainDir="/home/jeroen/test_package/",clustertype="SOCK",numCores=2)
}
}
\author{
Jeroen Pullens

adopted from Rconect package
}
\keyword{NUCOMBog}

