\name{runnucom_subset}
\alias{runnucom_subset}
\title{runnucom_subset}
\usage{
runnucom_subset(WD,par,type=NULL)
}
\arguments{
\item{WD} {"Working directory"}
\item{par} {"Parameter values (list of 26 values)"}
\item{type} {"Type to data to be retrieved, see getData()"}

}
\value{
run the model, with a subset of parameters (26)
}
\description{
run the model, with a subset of parameters (26)
}
\examples{
runnucom_subset(WD,par,type=NULL)
}
