% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runnucom_wrapper.R
\name{runnucom_wrapper}
\alias{runnucom_wrapper}
\title{Run NUCOMBog wrapper}
\usage{
runnucom_wrapper(runParameters)
}
\arguments{
\item{runParameters}{list of:
       $WD = Working directory
       $climate = name of climate file
       $environment = name of environment file
       $inival = name of initial values file
       $start  = start year
       $end = end year
       $par = data.frame(names,values)}
}
\description{
Code to run NUCOMBog wrapper, escpecially made for the parallel computing
}
\author{
Jeroen Pullens
}
\keyword{NUCOMBog}

