% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setupParallel.r
\name{setupParallel}
\alias{setupParallel}
\title{make setupParallel}
\usage{
setupParallel(mainDir, climate, environment, inival, start, end, type,
  parameters)
}
\arguments{
\item{climate}{climate input (monthly) format: year month temp prec pot evap (tab seperated)}

\item{environment}{environment input (yearly) format: year co2 nitrogen deposition}

\item{inival}{initial values of biomass}

\item{start}{year in which to start}

\item{end}{year in which to end}

\item{type}{which output do you want? "NEE" and/or "WTD"}

\item{parameters}{possibility to add own parameters, BUT is has to in a datafram format with colum names: "names" and "values", see example.}

\item{WD}{Working directory}
}
\value{
a list with paths and filenames and parameter values which can be implemented in runnucomParallel
}
\description{
Code to make setupParallel, this is needed to run the model in parallel.
}
\examples{
\dontrun{
names<-c("CO2ref","gram_Beta","sd1","sd2")

initialParameters<-list()
nparvector<-50
for (i in 1:nparvector) initialParameters[[i]]<-data.frame(names,values=runif(n=length(names),min=c(300,0.1,0.01,0.01),max=c(500,1,1,1)))
test_setup<-setupParallel(mainDir="/home/jeroen/test_package/",climate="clim_1999-2013_measured.txt",environment="Env_Mer_Bleue_1999_2013.txt",inival="Inival_Mer_Bleue.txt",start=1999,end=2013,type=c("NEE","WTD"),parameters=initialParameters)
}
}
\author{
Jeroen Pullens
}
\keyword{NUCOMBog}

