% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Function to retrieve data from the monthly output file created by NUCOMBog}
\usage{
getData(WD, type)
}
\arguments{
\item{WD}{Working Directory}

\item{type}{Data to be retrieved, integrated are "NEE" and "WTD"

The model gives NPP back as output, the model has been changed to give Decomposition as output
The model gives WTD in mm and negative means below ground level. by dividing it by 1000 the output is in meters and converted, that positive values mean water table depth.}
}
\description{
Function to retrieve data from the monthly output file created by NUCOMBog
}
\examples{
\dontrun{
getData(WD="~/nucom",type = c("NEE","WTD"))
}
}
\author{
Jeroen Pullens
}

