% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Function to retrieve data from the monthly output file created by NUCOMBog}
\usage{
getData(setup, startval)
}
\arguments{
\item{setup}{setup_structure described in setup_NUCOM}

\item{startval}{When a spinup is used for the model, not all output is needed. This "startval" parameter can be used to cut the output off, i.e. the row from which the "outmo.txt" file is loaded.}
}
\description{
This function returns data from the monthly output file created by NUCOMBog.

The original model gives NPP back as output, the model has been changed to give Autotrophic respiration as output. In this way NEE can be calculated.
The model gives WTD in meters and a positive value means below ground level.

This getData function is integrated in all runnucom functions.
}
\examples{
\dontrun{
getData(setup=test_setup_singlecore,startval=1)
}
}
\author{
JWM Pullens
}

