% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_filenames.r
\name{make_filenames}
\alias{make_filenames}
\title{make_filenames}
\usage{
make_filenames(WD, climate, environment, inival, start, end)
}
\arguments{
\item{WD}{Working Directory}

\item{climate}{climate input (monthly) format: year month temp prec pot evap (tab seperated)}

\item{environment}{environment input (yearly) format: year co2 nitrogen deposition}

\item{inival}{initial values of biomass}

\item{start}{year in which to start}

\item{end}{year in which to end}
}
\description{
make_filenames
}
\author{
Jeroen Pullens
}

