% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_param_file_complete.r
\name{make_param_file}
\alias{make_param_file}
\title{make_param_file_complete}
\usage{
make_param_file(WD, x)
}
\arguments{
\item{WD}{Working Directory}

\item{x}{values for the parameters}
}
\description{
make_param_file_complete
}
\examples{
\dontrun{
make_param_file(WD,c(1:198))
}
}
\author{
jeroen pullens
}

