\name{make_param_file_subset}
\alias{make_param_file_subset}
\title{make_param_file_subset}
\usage{
make_param_file_subset(WD,x)
}
\arguments{
\item{WD} {"Working directory"}
\item{x} {"Parameter values (list of 26 values)"}

}
\value{
Makes a parameter file to start the model
}
\description{
Makes a parameter file to start the model
}
\examples{
make_param_file_subset(WD,x)
}
