% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runnucom.r, R/runnucomParallel.r
\name{runnucom}
\alias{runnucom}
\title{runnucom
make function to run modelMEE}
\usage{
runnucom(WD, par, subset = F, type = NULL, parallel = F)

runnucom(WD, par, subset = F, type = NULL, parallel = F)
}
\arguments{
\item{WD}{Working Directory}

\item{par}{parameter list}

\item{subset}{Is there a subset of the parameters used?}

\item{Type}{to data to be retrieved, see getData()}

\item{WD}{}
}
\description{
runnucom
make function to run modelMEE

make function to run modelMEE ----
}
\examples{
\dontrun{
runnucom(WD,par, subset = F, type=NULL)
}
}
\author{
jeroen pullens

jeroen pullens
}

