\name{runnucom}
\alias{runnucom}
\title{runnucom}
\usage{
runnucom(WD,par,type=NULL)
}
\arguments{
\item{WD} {"Working directory"}
\item{par} {"Parameter values (list of 26 values)"}
\item{type} {"Type to data to be retrieved, see getData()"}

}
\value{
run the model
}
\description{
run the model
}
\examples{
runnucom(WD,par,type=NULL)
}
