% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runnucom.r
\name{runnucom}
\alias{runnucom}
\title{run NUCOMBog}
\usage{
runnucom(WD, climate, environment, inival, start, end, par = NULL,
  type = NULL)
}
\arguments{
\item{WD}{Working Directory}

\item{climate}{climate input (monthly) format: year month temp prec pot evap (tab seperated)}

\item{environment}{environment input (yearly) format: year co2 nitrogen deposition}

\item{inival}{initial values of biomass}

\item{start}{year in which to start}

\item{end}{year in which to end}

\item{type}{which output do you want? "NEE" and/or "WTD"}
}
\description{
Code to run NUCOMBog.
}
\examples{
\dontrun{
runnucom(WD="/home/jeroen/test_package/",climate="ClimateLVM.txt",environment="EnvironmentLVM.txt",inival="inivalLVM.txt",start=1800,end=1805,type=c("NEE","WTD"))
}
}
\author{
Jeroen Pullens
}

