% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runnucomParallel.r
\name{runnucom_parallel}
\alias{runnucom_parallel}
\title{Run NUCOMBog_parallel}
\usage{
runnucom_parallel(setup, clustertype, numCores = 1, parameters)
}
\arguments{
\item{setup}{The setup needs to be made before by running the setup_NUCOM function.}

\item{clustertype}{Clustertype: "MPI" is cluster and personal computer is "SOCK"}

\item{numCores}{Number of Cores on which are model needs to be run (NOTE: Non-parallel runs can only be run on 1 core)}

\item{parameters}{List of parameters}
}
\description{
Code to run NUCOMBog parallel, but first run setup_NUCOM
}
\examples{
\dontrun{
!!test_setup is from the function setup_NUCOM!!

parallel<-runnucom_parallel(setup = test_setup,
                            clustertype = "SOCK",
                            numCores = 1,
                            parameters=initialParameters)
}
}
\author{
JWM Pullens
}
\keyword{NUCOMBog}

