% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setupParallel.r
\name{setupParallel}
\alias{setupParallel}
\title{make setupParallel}
\usage{
setupParallel(WD, climate, environment, inival, start, end, type, parameters)
}
\arguments{
\item{WD}{Working directory}

\item{climate}{climate input (monthly) format: year month temp prec pot evap (tab seperated)}

\item{environment}{environment input (yearly) format: year co2 nitrogen deposition}

\item{inival}{initial values of biomass}

\item{start}{year in which to start}

\item{end}{year in which to end}

\item{type}{which output do you want? "NEE" and/or "WTD"}

\item{parameters}{possibility to add own parameters, BUT is has to in a datafram format with colum names: "names" and "values", see example.}
}
\value{
a list with paths and filenames and parameter values which can be implemented in runnucomParallel
}
\description{
Code to make setupParallel, this is needed to run the model in parallel.
}
\examples{
\dontrun{
names<-rep(c("CO2ref","gram_Beta"),each=50)
values<-c(seq(300,500,length.out = 50),seq(0.1,1,length.out = 50))
test_par<-data.frame(names,values)
test_setup<-setupParallel(WD="/home/jeroen/test_package/",climate="ClimateLVM.txt",environment="EnvironmentLVM.txt",inival="inivalLVM.txt",start=1800,end=1805,type=c("NEE","WTD"),parameters=test_par)
}
}
\author{
Jeroen Pullens
}
\keyword{NUCOMBog}

